@@ -40,17 +40,17 @@
         "registry": "https://repository.hybris.com/api/npm/npm-repository/"
     },
     "dependencies": {
-        "@angular/animations": "20.1.4",
-        "@angular/cdk": "20.1.4",
-        "@angular/common": "20.1.4",
-        "@angular/compiler": "20.1.4",
-        "@angular/core": "20.1.4",
-        "@angular/forms": "20.1.4",
-        "@angular/platform-browser": "20.1.4",
-        "@angular/platform-browser-dynamic": "20.1.4",
-        "@angular/router": "20.1.4",
+        "@angular/animations": "21.1.0",
+        "@angular/cdk": "21.1.0",
+        "@angular/common": "21.1.0",
+        "@angular/compiler": "21.1.0",
+        "@angular/core": "21.1.0",
+        "@angular/forms": "21.1.0",
+        "@angular/platform-browser": "21.1.0",
+        "@angular/platform-browser-dynamic": "21.1.0",
+        "@angular/router": "21.1.0",
         "@fundamental-styles/cx": "0.40.1",
-        "@nx/angular": "21.2.4",
+        "@nx/angular": "22.3.3",
         "@sap-theming/theming-base-content": "11.32.3",
         "@stackblitz/sdk": "1.9.0",
         "@swc/helpers": "0.5.11",
@@ -152,12 +152,12 @@
         "semver": "7.6.3",
         "standard-version": "9.5.0",
         "through2": "4.0.2",
-        "ts-jest": "29.4.0",
+        "ts-jest": "29.4.6",
         "ts-node": "10.9.2",
         "tsconfig-paths": "4.1.0",
         "typedoc": "0.25.3",
         "typedoc-plugin-merge-modules": "5.1.0",
-        "typescript": "5.8.2",
+        "typescript": "5.9.3",
         "typescript-eslint": "^8.33.1",
         "wait-on": "6.0.1",
         "wdio-axe": "2.2.3",
