@@ -5,14 +5,13 @@ import {
     DOCUMENT,
     Directive,
     ElementRef,
-    HostBinding,
-    HostListener,
-    Input,
     OnDestroy,
     PLATFORM_ID,
     Renderer2,
     booleanAttribute,
-    inject
+    computed,
+    inject,
+    input
 } from '@angular/core';
 import { HasElementRef } from '@fundamental-ngx/cdk/utils';
 import scrollbarStyles from 'fundamental-styles/dist/js/scrollbar';
@@ -37,49 +36,38 @@ let styleSheet: HTMLStyleElement | null = null;
 @Directive({
     selector: '[fdScrollbar], [fd-scrollbar]',
     host: {
-        class: 'fd-scrollbar'
+        class: 'fd-scrollbar',
+        '[style.overflow-x]': '_overflowX()',
+        '[style.overflow-y]': '_overflowY()',
+        '(scroll)': 'onScroll($event)'
     },
-    hostDirectives: [CdkScrollable],
-    standalone: true
+    hostDirectives: [CdkScrollable]
 })
 export class ScrollbarDirective implements OnDestroy, HasElementRef {
     /** Whether overflow horizontal content should be hidden. */
-    @Input({ transform: booleanAttribute })
-    noHorizontalScroll = false;
+    readonly noHorizontalScroll = input(false, { transform: booleanAttribute });
 
     /** Whether overflow vertical content should be hidden. */
-    @Input({ transform: booleanAttribute })
-    noVerticalScroll = false;
+    readonly noVerticalScroll = input(false, { transform: booleanAttribute });
 
     /** Whether scrollbars should be visible even if content fits. */
-    @Input({ transform: booleanAttribute })
-    alwaysVisible = false;
+    readonly alwaysVisible = input(false, { transform: booleanAttribute });
 
-    /** @hidden */
-    @HostBinding('style.overflow-x')
-    get _overflowX(): ScrollbarOverflowOptions {
-        if (this.noHorizontalScroll) {
-            return 'hidden';
-        }
-
-        return this._overflow;
-    }
-
-    /** @hidden */
-    @HostBinding('style.overflow-y')
-    get _overflowY(): ScrollbarOverflowOptions {
-        if (this.noVerticalScroll) {
-            return 'hidden';
-        }
-
-        return this._overflow;
-    }
-
-    /** @hidden */
+    /** @hidden Whether the scrollbar is inside a popover and should stop scroll event propagation. */
     _inPopover = false;
 
     /** @hidden */
     elementRef: ElementRef<HTMLElement> = inject(ElementRef);
+
+    /** @hidden */
+    protected readonly _overflowX = computed<ScrollbarOverflowOptions>(() =>
+        this.noHorizontalScroll() ? 'hidden' : this.alwaysVisible() ? 'scroll' : 'auto'
+    );
+
+    /** @hidden */
+    protected readonly _overflowY = computed<ScrollbarOverflowOptions>(() =>
+        this.noVerticalScroll() ? 'hidden' : this.alwaysVisible() ? 'scroll' : 'auto'
+    );
 
     /** @hidden */
     private _document: Document = inject(DOCUMENT);
@@ -106,7 +94,6 @@ export class ScrollbarDirective implements OnDestroy, HasElementRef {
     }
 
     /** @hidden */
-    @HostListener('scroll', ['$event'])
     onScroll(event: Event): void {
         if (this._inPopover) {
             event.stopImmediatePropagation();
@@ -124,14 +111,5 @@ export class ScrollbarDirective implements OnDestroy, HasElementRef {
     /** method to invoke scroll */
     scroll(options: ScrollToOptions): void {
         this.elementRef.nativeElement.scroll(options);
-    }
-
-    /** @hidden */
-    private get _overflow(): ScrollbarOverflowOptions {
-        if (this.alwaysVisible) {
-            return 'scroll';
-        }
-
-        return 'auto';
     }
 }
