@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { ButtonComponent } from '@fundamental-ngx/core/button';
-import { ScrollbarModule } from '@fundamental-ngx/core/scrollbar';
+import { ScrollbarDirective } from '@fundamental-ngx/core/scrollbar';
 import { I18nModule } from '@fundamental-ngx/i18n';
 import { SideNavigationButtonDirective } from './side-navigation-button.directive';
 import { SideNavigationMainComponent } from './side-navigation-main.component';
@@ -14,7 +14,7 @@ import { SideNavigationComponent } from './side-navigation.component';
 @NgModule({
     imports: [
         ButtonComponent,
-        ScrollbarModule,
+        ScrollbarDirective,
         I18nModule,
         SideNavigationMainComponent,
         SideNavigationUtilityDirective,
