@@ -3,7 +3,7 @@ import { LiveAnnouncer } from '@angular/cdk/a11y';
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, DestroyRef } from '@angular/core';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { BusyIndicatorComponent } from '@fundamental-ngx/core/busy-indicator';
-import { InfiniteScrollModule } from '@fundamental-ngx/core/infinite-scroll';
+import { InfiniteScrollDirective } from '@fundamental-ngx/core/infinite-scroll';
 import { ListModule } from '@fundamental-ngx/core/list';
 import { of } from 'rxjs';
 import { delay } from 'rxjs/operators';
@@ -14,7 +14,7 @@ const ITEMS_AMOUNT_ON_LOAD = 5;
     selector: 'fd-list-infinite-scroll-example',
     templateUrl: './list-infinite-scroll-example.component.html',
     changeDetection: ChangeDetectionStrategy.OnPush,
-    imports: [BusyIndicatorComponent, ListModule, InfiniteScrollModule]
+    imports: [BusyIndicatorComponent, ListModule, InfiniteScrollDirective]
 })
 export class ListInfiniteScrollExampleComponent {
     // List that is displayed to the user
