@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { ButtonComponent } from '@fundamental-ngx/core/button';
 import { ContentDensityModule } from '@fundamental-ngx/core/content-density';
-import { ScrollbarModule } from '@fundamental-ngx/core/scrollbar';
+import { ScrollbarDirective } from '@fundamental-ngx/core/scrollbar';
 import {
     NestedItemComponent,
     NestedLinkComponent,
@@ -30,7 +30,7 @@ import { I18nModule } from '@fundamental-ngx/i18n';
     standalone: true,
     imports: [
         ButtonComponent,
-        ScrollbarModule,
+        ScrollbarDirective,
         I18nModule,
         SideNavigationMainComponent,
         SideNavigationUtilityDirective,
