@@ -10,7 +10,7 @@ import { BarModule } from '@fundamental-ngx/core/bar';
 import { ContentDensityDirective } from '@fundamental-ngx/core/content-density';
 import { DialogModule } from '@fundamental-ngx/core/dialog';
 import { FormLabelComponent } from '@fundamental-ngx/core/form';
-import { LayoutGridModule } from '@fundamental-ngx/core/layout-grid';
+import { LayoutGridColDirective, LayoutGridComponent, LayoutGridRowDirective } from '@fundamental-ngx/core/layout-grid';
 import { ScrollbarDirective } from '@fundamental-ngx/core/scrollbar';
 import { TitleComponent } from '@fundamental-ngx/core/title';
 import { PlatformButtonModule } from '@fundamental-ngx/platform/button';
@@ -30,7 +30,9 @@ import {
         PlatformButtonModule,
         ContentDensityDirective,
         TitleComponent,
-        LayoutGridModule,
+        LayoutGridColDirective,
+        LayoutGridComponent,
+        LayoutGridRowDirective,
         FormLabelComponent,
         PlatformLinkModule,
         DialogModule,
